Refactor decor entry config
