+++ Assets/Scripts/MapGeneration/MapGenerationConfig.cs
@@ -69,8 +69,8 @@ namespace TimelessEchoes.MapGeneration
             [HideInInspector] [TabGroup("Decor", "References")] public Tilemap decorMap;
             [Range(0f, 1f)] public float density = 1f;
             [Searchable]
-            [ListDrawerSettings(Expanded = false)]
-            [TabGroup("Decor", "Items")] public List<TilemapChunkGenerator.DecorEntry> decor = new();
+            [ListDrawerSettings(ShowFoldout = true, DefaultExpandedState = false)]
+            [TabGroup("Decor", "Items")] public List<DecorEntry> decor = new();
         }
     }
 }
